Stock et livraison : lire ce que le serveur dit vraiment
Volet `lilia-food-admin` de la remédiation de l'audit du 05/09/2026
(AUDIT_2026-09-05_livraison_stock.md).

## `isAvailable` était masqué par un getter

    bool get isAvailable => stockRestant == null || stockRestant! > 0;  // ✗

Ce getter portait le nom d'un champ du serveur et le recouvrait : la
disponibilité déclarée par le vendeur n'était **jamais lue**. Deux notions que
le backend avait explicitement séparées en août 2026 (fix M2) — « retiré de la
vente » est une décision, « épuisé » une conséquence — étaient refusionnées
côté client.

La conséquence est réelle ici, contrairement à l'app cliente : cet écran lit
`GET /products/manage`, qui montre **volontairement** les produits retirés. Un
produit que le vendeur venait de retirer s'affichait « Stock: 4/10 », en bleu,
sur l'écran même qui porte le bouton pour le remettre en vente.

Le getter devient `isInStock`, le champ serveur reprend son nom, et la question
que posent les écrans a désormais un seul endroit : `isOrderable = isAvailable
&& isInStock`.

## Un champ qui lisait une clé inexistante

Le modèle déclarait `isActive`, lu depuis `json['isActive']` — **une clé qui
n'existe pas** sur `Product` côté serveur : le modèle Prisma porte
`isAvailable` et `deletedAt`, jamais `isActive`. Il valait donc toujours `true`
par le défaut du parseur, et traversait `copyWith` sans que personne s'en
serve. Vérifié contre la production : la réponse ne porte que `isAvailable`.
Le champ fantôme **est** devenu le vrai `isAvailable`.

## Deux routes n'avaient aucun appelant

- `PATCH /products/:id/availability` — le seul levier du vendeur pour retirer
  un produit était de mettre son stock à 0, ce qui affiche « épuisé » : une
  information différente, et fausse. Afficher « Retiré de la vente » sans
  offrir le retour aurait été un demi-correctif.
- `PATCH /products/:id/stock` — la seule route qui remette `stockRestant` à
  niveau. Le formulaire produit décrit la **capacité** ; réapprovisionner est
  un autre geste, et les confondre était la moitié du bug S-1 (côté serveur,
  `stockQuotidien` était absent du DTO de mise à jour, donc supprimé en silence
  par le `ValidationPipe`).

## Écran de livraison

`PATCH /vendors/:id/delivery` remplace `PATCH /restaurants/:id/delivery-settings`,
dépréciée. Les deux écrivent les mêmes colonnes, mais l'ancienne validait
moins : `deliveryPriceMode` n'y était qu'un `@IsString()` — « GRATUIT » passait
et remontait en 500 — et les montants n'avaient pas de borne haute. Deux portes
sur la même chose finissent toujours par ne pas poser les mêmes règles.

⚠️ Les montants passent de `double` à `int` : le XAF n'a pas de sous-unité, et
le serveur exige `@IsInt` depuis la migration `money_integers`. Un tarif saisi
« 1500,5 » partait en 400.

`flutter analyze` : 0 erreur, 0 warning. Tests : 115/115.

+++ lib/features/settings/presentation/providers/settings_provider.dart
@@ -60,11 +60,14 @@ class RestaurantSettings extends _$RestaurantSettings {
     });
   }
 
+  /// Montants en `int` : le XAF n'a pas de sous-unité, et le serveur les
+  /// valide en `@IsInt` depuis la migration `money_integers`. Un `double`
+  /// portant une décimale se faisait refuser en 400.
   Future<void> updateDeliverySettings({
-    double? fixedDeliveryFee,
+    int? fixedDeliveryFee,
     int? estimatedDeliveryTimeMin,
     int? estimatedDeliveryTimeMax,
-    double? minimumOrderAmount,
+    int? minimumOrderAmount,
     String? deliveryPriceMode,
   }) async {
     final currentRestaurant = state.value;
